# docs/api-spec.yml
...
openapi: 3.0.3
info:
  title: Share Social Media API
  version: "1.0.0"
  description: >
    Express API for the Share Social Media demo. Consumed by the Vite React SPA.
    Persistence is DynamoDB single-table; media is S3 (+ CloudFront in AWS).
    Authentication is dual-mode (local HS256 JWT vs Cognito access tokens).
  contact:
    name: Sebastian Gonzalez
    url: https://github.com/JuanSebastianGB/share-social-media
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development (server.ts)
  - url: https://{apiId}.execute-api.{region}.amazonaws.com
    description: AWS HTTP API (Lambda). Replace from stack output ApiUrl.
    variables:
      apiId:
        default: example
      region:
        default: us-east-1
...
paths:
...
  /users/{id}/posts:
    get:
      tags: [users]
      summary: List posts authored by user
      operationId: getUserPosts
      parameters:
        - $ref: "#/components/parameters/MongoIdPath"
      responses:
        "200":
          description: Hydrated posts for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HydratedPost"
        "403":
          description: Validation or error string
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
...
  /posts:
    get:
      tags: [posts]
      summary: Paginated feed
      description: >
        Query page (default 1) and optional search. Server hard-codes limit=2.
        Returns hydrated posts matching search against body and user fields.
      operationId: getPostsPagination
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Page of hydrated posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HydratedPost"
    post:
      tags: [posts]
      summary: Create text post (no new upload)
      description: Requires JWT. Uses validatorCreatePost (body, type).
      operationId: createUserPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "200":
          description: Created / resulting post payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HydratedPost"
        "403":
          description: Auth, validation, or ERROR_*
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
  /posts/file:
    post:
      tags: [posts]
      summary: Create post with image upload
      description: JWT + multer myFile + S3 upload middleware + createUserPostFile.
      operationId: createUserPostFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [body, myFile]
              properties:
                body:
                  type: string
                type:
                  type: string
                myFile:
                  type: string
                  format: binary
      responses:
        "200":
          description: Created hydrated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HydratedPost"
        "403":
          description: Missing file, validation, or ERROR_*
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
  /posts/{id}:
    get:
      tags: [posts]
      summary: Get post by id
      operationId: getPost
      parameters:
        - $ref: "#/components/parameters/MongoIdPath"
      responses:
        "200":
          description: Array with one hydrated post (controller returns getPostService array)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HydratedPost"
        "403":
          description: Validation or ERROR_GET_POST
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
    put:
      tags: [posts]
      summary: Toggle like on post
      operationId: toggleLikePost
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MongoIdPath"
      responses:
        "200":
          description: Updated post / likes payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "403":
          description: Auth, validation, or ERROR_*
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
    delete:
      tags: [posts]
      summary: Delete post (and hard-delete media when present)
      operationId: deletePost
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MongoIdPath"
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Auth, validation, non-owner ERROR_NOT_RESOURCE_OWNER, or ERROR_*
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
  /posts/{id}/comments:
    get:
      tags: [posts]
      summary: List comments for a post
      operationId: getPostComments
      parameters:
        - $ref: "#/components/parameters/MongoIdPath"
      responses:
        "200":
          description: Comments resolved from post.comments ids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "403":
          description: Validation or error
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorString"
                  - $ref: "#/components/schemas/ValidationErrorBody"
...
components:
  parameters:
    MongoIdPath:
      name: id
      in: path
      required: true
      description: 24-char hex id (express-validator isMongoId)
      schema:
        type: string
        pattern: "^[a-fA-F0-9]{24}$"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Local mode — HS256 JWT from /auth/login signed with JWT_SECRET.
        Cognito mode — Cognito access token verified with aws-jwt-verify.
  schemas:
    ErrorString:
      description: handleHttpErrors body — a JSON-encoded string, not an object
      type: string
      example: ERROR_LOGIN
    ValidationErrorBody:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
            description: express-validator error object
    User:
      type: object
      required: [_id, email, friends, role]
      properties:
        _id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        password:
          type: string
          description: Omitted or undefined on safe responses
        email:
          type: string
          format: email
        age:
          type: number
        role:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        friends:
          type: array
          items:
            type: string
        location:
          type: string
        occupation:
          type: string
        viewedProfile:
          type: number
        impressions:
          type: number
        profileImageId:
          type: string
        cognitoSub:
          type: string
        profileImage:
          $ref: "#/components/schemas/File"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    File:
      type: object
      required: [_id]
      properties:
        _id:
          type: string
        fileName:
          type: string
        filename:
          type: string
        url:
          type: string
        deleted:
          type: boolean
        userId:
          type: string
          description: Optional owner USER `_id` (domain `ownerId`). Not set on ownerless rows.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Post:
      type: object
      required: [_id, body, userId, likes, comments]
      properties:
        _id:
          type: string
        body:
          type: string
        userId:
          type: string
        fileId:
          type: string
        likes:
          type: object
          additionalProperties:
            type: boolean
          description: Map of userId to true when liked
        comments:
          type: array
          items:
            type: string
          description: Comment ids
        type:
          type: string
          default: file
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HydratedPost:
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            file:
              type: object
              properties:
                _id:
                  type: string
                url:
                  type: string
            user:
              $ref: "#/components/schemas/User"
    CreatePostBody:
      type: object
      required: [body]
      properties:
        body:
          type: string
        type:
          type: string
    Comment:
      type: object
      required: [_id, description, userId, firstName, lastName]
      properties:
        _id:
          type: string
        description:
          type: string
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DeleteResult:
      type: object
      required: [acknowledged, deletedCount]
      properties:
        acknowledged:
          type: boolean
          enum: [true]
        deletedCount:
          type: integer
...